Start
Included the "Gridly" plugin. Plug in is jquery based and is found on http://ksylvest.github.io/jquery-gridly/ I feel this is the right way to go. Gridly supports reorganization for the drag and drop and is supported on mobile. Also uses differently sized boxes as we want them to.